requirements.txt
langchain==0.3.1
langgraph==0.2.28
langchain_openai==0.2.2
streamlit==1.39.0
openai==1.51.2
python-dotenv==1.0.1
supabase==2.27.1